requirements.txt
wheel==0.44.0
# https://github.com/ronaldoussoren/py2app/issues/538#issuecomment-3193389958
# Don't upgrade past this version as 70.x.x does not work
# setuptools==70.1.1
setuptools==70.3.0

# the following necessary when building the
# application with py2App
chardet==5.2.0

py2app==0.28.9

mypy==1.18.2
html-testRunner~=1.2.1
pyautogui==0.9.54
pillow==12.0.0
PyMsgBox==2.0.1
# So Circle CI works

buildlackey==2.0.0

wxPython==4.2.4
Pypubsub==4.0.3
semantic-version==2.10.0
codeallybasic==1.30.0
codeallyadvanced==2.3.0
umlmodel==3.4.0

umlshapes==1.5.0
umlio==0.5.3
umlextensions==0.3.1
